+++ .github/workflows/publish.yml
@@ -1,50 +1,56 @@
 name: Release and Publish
 
 on:
- push:
-  branches: 
-    - master
-    - alpha
-    - beta
-    - next
+  push:
+    branches:
+      - master
+      - alpha
+      - beta
+      - next
 
 jobs:
- test:
-  runs-on: ubuntu-latest
-  steps:
-  - name: Checkout
-    uses: actions/checkout@v4
-  - name: Setup Node.js
-    uses: actions/setup-node@v4
-    with:
-      node-version: 20.x
-  - name: npm install
-    run: |
-      npm install
-  - name: npm lint
-    run: |
-      npm run lint
-  - name: npm test
-    run: |
-      npm test
+  test:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+      - name: Setup Node.js
+        uses: actions/setup-node@v4
+        with:
+          node-version: 20.x
 
- release:
-  name: Release
-  runs-on: ubuntu-latest
-  needs: [test]
-  steps:
-  - name: Checkout
-    uses: actions/checkout@v4
-  - name: Setup Node.js
-    uses: actions/setup-node@v4
-    with:
-      node-version: 20.x
-  - name: npm install
-    run: |
-      npm install
-  - name: npx semantic-release
-    run: |
-      npx semantic-release
-    env:
-      GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-      NPM_TOKEN: ${{ secrets.NPM_TOKEN }}
+      - name: npm install
+        run: npm install
+
+      - name: npm lint
+        run: npm run lint
+
+      - name: npm types
+        run: npm run types
+
+      - name: npm test
+        run: npm test
+
+  release:
+    name: Release
+    runs-on: ubuntu-latest
+    needs: [test]
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+      - name: Setup Node.js
+        uses: actions/setup-node@v4
+        with:
+          node-version: 20.x
+
+      - name: npm install
+        run: npm install
+
+      - name: npm types
+        run: npm run types
+
+      - name: npx semantic-release
+        run: npx semantic-release
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+          NPM_TOKEN: ${{ secrets.NPM_TOKEN }}
