+++ .prettierrc
@@ -1,5 +1,19 @@
 {
     "singleQuote": true,
     "trailingComma": "all",
-    "tabWidth": 4
+    "tabWidth": 4,
+    "overrides": [
+        {
+            "files": [
+                ".prettierrc",
+                "*.json",
+                "*.yml",
+                ".travis.yml",
+                ".eslintrc"
+            ],
+            "options": {
+                "tabWidth": 2
+            }
+        }
+    ]
 }
