+++ lib/asset-css.js
@@ -6,11 +6,34 @@ const inspect = Symbol.for('nodejs.util.inspect.custom');
 // Ref: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/link
 // NOTE: Only includes attributes used for loading CSS
 
-const toUndefined = value => {
+/**
+ * Returns the value, or undefined if the value is false or an empty string.
+ * @template T
+ * @param {T} value
+ * @returns {T | undefined}
+ */
+const toUndefined = (value) => {
     if (value === false) return undefined;
     if (value === '') return undefined;
     return value;
 };
+
+/**
+ * @typedef {object} CssAsset
+ * @property {string} value
+ * @property {string} [type]
+ * @property {string} [rel]
+ * @property {string | boolean} [crossorigin]
+ * @property {string} [hreflang=""]
+ * @property {boolean} [prefix=false]
+ * @property {boolean} [disabled=false]
+ * @property {string} [pathname=""]
+ * @property {string} [title=""]
+ * @property {string} [media=""]
+ * @property {string} [as=""]
+ * @property {string} [strategy]
+ * @property {string} [scope]
+ */
 
 export default class PodiumAssetCss {
     #crossorigin;
@@ -26,6 +49,24 @@ export default class PodiumAssetCss {
     #as;
     #strategy;
     #scope;
+
+    /**
+     * @constructor
+     * @param {object} options
+     * @param {boolean | string | null} [options.crossorigin]
+     * @param {string | null} [options.hreflang=""]
+     * @param {boolean | null} [options.prefix=false]
+     * @param {boolean | null} [options.disabled=false]
+     * @param {string | null} [options.pathname=""]
+     * @param {string | null} [options.title=""]
+     * @param {string | null} [options.media=""]
+     * @param {string | null} [options.type="text/css"]
+     * @param {string | null} [options.rel="stylesheet"]
+     * @param {string | null} [options.as=""]
+     * @param {string | null} [options.value]
+     * @param {string | null} [options.strategy]
+     * @param {string | null} [options.scope]
+     */
     constructor({
         crossorigin = undefined,
         pathname = '',
@@ -160,6 +201,9 @@ export default class PodiumAssetCss {
         this.#scope = value;
     }
 
+    /**
+     * @returns {CssAsset}
+     */
     toJSON() {
         return {
             crossorigin: toUndefined(this.crossorigin),
@@ -172,7 +216,7 @@ export default class PodiumAssetCss {
             rel: this.rel,
             as: toUndefined(this.as),
             strategy: toUndefined(this.strategy),
-            scope: toUndefined(this.scope)
+            scope: toUndefined(this.scope),
         };
     }
 
@@ -193,7 +237,7 @@ export default class PodiumAssetCss {
             as: this.as,
         };
     }
-    
+
     toJsxAttributes() {
         return buildReactLinkAttributes(this);
     }
@@ -201,4 +245,4 @@ export default class PodiumAssetCss {
     get [Symbol.toStringTag]() {
         return 'PodiumAssetCss';
     }
-};
+}
