+++ lib/asset-js.js
@@ -9,12 +9,33 @@ const inspect = Symbol.for('nodejs.util.inspect.custom');
 // Ref: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/script
 // NOTE: "nonce" is deliberately left out since we do not support inline scripts
 
+/**
+ * Returns the value, or undefined if the value is false, an empty array or an empty string.
+ * @template T
+ * @param {T} value
+ * @returns {T | undefined}
+ */
 const toUndefined = (value) => {
     if (Array.isArray(value) && value.length === 0) return undefined;
     if (value === false) return undefined;
     if (value === '') return undefined;
     return value;
 };
+
+/**
+ * @typedef {object} JavaScriptAsset
+ * @property {string} [referrerpolicy]
+ * @property {string | boolean} [crossorigin]
+ * @property {string} [integrity]
+ * @property {boolean} [nomodule]
+ * @property {string} value
+ * @property {boolean} [async]
+ * @property {boolean} [defer]
+ * @property {string} [type]
+ * @property {Array<{ key: string; value?: unknown }>} [data]
+ * @property {string} [strategy]
+ * @property {string} [scope]
+ */
 
 export default class PodiumAssetJs {
     #referrerpolicy;
@@ -32,20 +53,21 @@ export default class PodiumAssetJs {
     #scope;
 
     /**
+     * @constructor
      * @param {object} options
-     * @param {string} [options.referrerpolicy]
-     * @param {string} [options.crossorigin]
-     * @param {string} [options.integrity]
-     * @param {string} [options.pathname]
-     * @param {boolean} [options.nomodule=false]
-     * @param {boolean} [options.prefix=false]
-     * @param {string} [options.value]
-     * @param {boolean} [options.async=false]
-     * @param {boolean} [options.defer=false]
-     * @param {string} [options.type="default"]
+     * @param {string | null} [options.referrerpolicy]
+     * @param {boolean | string | null} [options.crossorigin]
+     * @param {string | null} [options.integrity]
+     * @param {string | null} [options.pathname]
+     * @param {boolean | null} [options.nomodule=false]
+     * @param {boolean | null} [options.prefix=false]
+     * @param {string | null} [options.value]
+     * @param {boolean | null} [options.async=false]
+     * @param {boolean | null} [options.defer=false]
+     * @param {string | null} [options.type="default"]
      * @param {array} [options.data]
-     * @param {string} [options.strategy]
-     * @param {string} [options.scope]
+     * @param {string | null} [options.strategy]
+     * @param {string | null} [options.scope]
      */
     constructor({
         referrerpolicy = '',
@@ -189,6 +211,9 @@ export default class PodiumAssetJs {
         this.#scope = value;
     }
 
+    /**
+     * @returns {JavaScriptAsset}
+     */
     toJSON() {
         return {
             referrerpolicy: toUndefined(this.referrerpolicy),
