+++ lib/html-utils.js
@@ -1,13 +1,21 @@
 /* eslint-disable no-restricted-syntax */
 
+/**
+ * @param {unknown} value
+ * @returns {boolean}
+ */
 export const notEmpty = (value) => {
-    if (value === false) return value;
+    if (value === false) return false;
     if (value === undefined) return false;
     if (value === null) return false;
     if (value !== '') return true;
     return false;
 };
 
+/**
+ * @param {import("./asset-js.js").JavaScriptAsset} obj
+ * @returns {Array<{ key: string; value?: unknown; }>}
+ */
 export const buildScriptAttributes = (obj) => {
     const args = [];
     // lazy uses dynamic import in the script body, everything else uses the src attribute
@@ -16,7 +24,11 @@ export const buildScriptAttributes = (obj) => {
     }
 
     // ESM and module are valid "module" types. Lazy requires type="module" so we set it here.
-    if (obj.type === 'esm' || obj.type === 'module' || obj.strategy === 'lazy') {
+    if (
+        obj.type === 'esm' ||
+        obj.type === 'module' ||
+        obj.strategy === 'lazy'
+    ) {
         args.push({ key: 'type', value: 'module' });
     }
 
@@ -54,6 +66,12 @@ export const buildScriptAttributes = (obj) => {
     return args;
 };
 
+/**
+ * Transforms HTML attribute names to JSX compatible camelCased props.
+ *
+ * @param {import("./asset-js.js").JavaScriptAsset} obj
+ * @returns {object}
+ */
 export const buildReactScriptAttributes = (obj) => {
     const attrs = {};
     for (const { key, value } of buildScriptAttributes(obj)) {
@@ -64,16 +82,24 @@ export const buildReactScriptAttributes = (obj) => {
         else attrs[key] = value;
     }
     return attrs;
-}
+};
 
+/**
+ * @param {import("./asset-js.js").JavaScriptAsset} obj
+ * @returns {string}
+ */
 export const buildScriptElement = (obj) => {
-    const attrs = buildScriptAttributes(obj).map(({key, value}) => {
+    const attrs = buildScriptAttributes(obj).map(({ key, value }) => {
         if (!value && value !== '') return key;
         return `${key}="${value}"`;
-    })
+    });
     return `<script ${attrs.join(' ')}>${obj.strategy === 'lazy' ? `import("${obj.value}");` : ''}</script>`;
 };
 
+/**
+ * @param {import("./asset-css.js").CssAsset} obj
+ * @returns {Array<{ key: string; value?: unknown; }>}
+ */
 export const buildLinkAttributes = (obj) => {
     const args = [];
     args.push({ key: 'href', value: obj.value });
@@ -114,6 +140,12 @@ export const buildLinkAttributes = (obj) => {
     return args;
 };
 
+/**
+ * Transforms HTML attribute names to JSX compatible camelCased props.
+ *
+ * @param {import("./asset-css.js").CssAsset} obj
+ * @returns {object}
+ */
 export const buildReactLinkAttributes = (obj) => {
     const attrs = {};
     for (const { key, value } of buildLinkAttributes(obj)) {
@@ -123,12 +155,16 @@ export const buildReactLinkAttributes = (obj) => {
         else attrs[key] = value;
     }
     return attrs;
-}
+};
 
+/**
+ * @param {import("./asset-css.js").CssAsset} obj
+ * @returns {string}
+ */
 export const buildLinkElement = (obj) => {
-    const attrs = buildLinkAttributes(obj).map(({key, value}) => {
+    const attrs = buildLinkAttributes(obj).map(({ key, value }) => {
         if (!value && value !== '') return key;
         return `${key}="${value}"`;
-    })
+    });
     return `<link ${attrs.join(' ')}>`;
 };
