+++ lib/http-incoming.js
@@ -2,12 +2,25 @@ import { URL } from 'node:url';
 
 const inspect = Symbol.for('nodejs.util.inspect.custom');
 
-const urlFromRequest = request => {
+const urlFromRequest = (request) => {
     const protocol = request?.protocol || 'http';
     const host = request?.headers?.host || 'localhost';
     const url = request?.url || '';
     return new URL(url, `${protocol.replace(/:/, '')}://${host}`);
 };
+
+/**
+ * @typedef {object} PodiumHttpIncoming
+ * @property {string} name
+ * @property {object} context
+ * @property {object} view
+ * @property {URL} url
+ * @property {Array<import('./asset-css.js').CssAsset | string>} css
+ * @property {Array<import('./asset-js.js').JavaScriptAsset | string>} js
+ * @property {object} [params]
+ * @property {boolean} [development]
+ * @property {boolean} [proxy]
+ */
 
 export default class HttpIncoming {
     #development;
@@ -18,9 +31,32 @@ export default class HttpIncoming {
     #proxy;
     #name;
     #view;
+
+    /**
+     * @type {URL | undefined}
+     */
     #url;
+
+    /**
+     * @type {Array<import('./asset-css.js').CssAsset | string>}
+     */
     #css;
+
+    /**
+     * @type {Array<import('./asset-js.js').JavaScriptAsset | string>}
+     */
     #js;
+
+    /**
+     * @param {object} [request={}] The incoming HTTP request
+     * @param {object} [response={}] The HTTP response
+     * @param {object} [params={}] Parameters such as locale. Typically res.locals.
+     *
+     * @example
+     * ```js
+     * const incoming = new HttpIncoming(req, res, res.locals);
+     * ```
+     */
     constructor(request = {}, response = {}, params = {}) {
         this.#development = false;
         this.#response = response;
@@ -33,7 +69,6 @@ export default class HttpIncoming {
         this.#url = undefined;
         this.#css = [];
         this.#js = [];
-
     }
 
     set development(value) {
@@ -44,6 +79,9 @@ export default class HttpIncoming {
         return this.#development;
     }
 
+    /**
+     * @throws {Error} Read only-property
+     */
     set response(value) {
         throw new Error('Cannot set read-only property.');
     }
@@ -68,24 +106,33 @@ export default class HttpIncoming {
         return this.#context;
     }
 
+    // No way to type writeonly properly at time of writing: https://github.com/microsoft/TypeScript/issues/21759
+    // This is a workaround for a type error where TS complains we're trying to set a value to something that is type void.
+    /**
+     * @type {unknown}
+     */
     set podlets(value) {
         const podlets = Array.isArray(value) ? value : [value];
 
-        podlets.forEach(podlet => {
+        podlets.forEach((podlet) => {
             if (podlet.css) {
-                podlet.css.forEach(item => {
+                podlet.css.forEach((item) => {
                     this.#css.push(item);
                 });
             }
 
             if (podlet.js) {
-                podlet.js.forEach(item => {
+                podlet.js.forEach((item) => {
                     this.#js.push(item);
                 });
             }
         });
     }
 
+    /**
+     * @type {unknown}
+     * @throws This prop is write only. Read from {@link css} and {@link js}.
+     */
     get podlets() {
         throw new Error(
             `The getter for .podlets is reserved for later implementation`,
@@ -130,7 +177,7 @@ export default class HttpIncoming {
 
     get url() {
         if (this.#url) return this.#url;
-        return urlFromRequest(this.#request);;
+        return urlFromRequest(this.#request);
     }
 
     set css(value) {
@@ -153,6 +200,9 @@ export default class HttpIncoming {
         return this.#js;
     }
 
+    /**
+     * @returns {PodiumHttpIncoming}
+     */
     toJSON() {
         return {
             development: this.development,
@@ -182,8 +232,8 @@ export default class HttpIncoming {
             js: this.js,
         };
     }
-    
+
     get [Symbol.toStringTag]() {
         return 'PodiumHttpIncoming';
     }
-};
+}
