+++ lib/main.js
@@ -1,15 +1,9 @@
-export { default as HttpIncoming } from './http-incoming.js';
-export { default as AssetCss } from './asset-css.js';
-export { default as AssetJs } from './asset-js.js';
-
-export { document as template } from './html-document.js';
-
-export { 
-    buildScriptElement, 
-    buildLinkElement, 
-} from './html-utils.js';
-
-export {
+import HttpIncoming from './http-incoming.js';
+import AssetCss from './asset-css.js';
+import AssetJs from './asset-js.js';
+import { document as template } from './html-document.js';
+import { buildScriptElement, buildLinkElement } from './html-utils.js';
+import {
     duplicateOnLocalsPodium,
     uriRelativeToAbsolute,
     getFromLocalsPodium,
@@ -22,3 +16,35 @@ export {
     isFunction,
     isString,
 } from './utils.js';
+
+/**
+ * @typedef {import('./asset-css.js').CssAsset} PodiumCssAsset
+ */
+
+/**
+ * @typedef {import('./asset-js.js').JavaScriptAsset} PodiumJavaScriptAsset
+ */
+
+/**
+ * @typedef {import('./http-incoming.js').PodiumHttpIncoming} PodiumHttpIncoming
+ */
+
+export {
+    HttpIncoming,
+    AssetCss,
+    AssetJs,
+    template,
+    buildLinkElement,
+    buildScriptElement,
+    duplicateOnLocalsPodium,
+    uriRelativeToAbsolute,
+    getFromLocalsPodium,
+    deserializeContext,
+    setAtLocalsPodium,
+    serializeContext,
+    pathnameBuilder,
+    uriIsRelative,
+    uriBuilder,
+    isFunction,
+    isString,
+};
