+++ package.json
@@ -1,60 +1,63 @@
 {
-    "name": "@podium/utils",
-    "version": "5.0.3",
-    "description": "Common generic utility methods shared by @podium modules.",
-    "type": "module",
-    "license": "MIT",
-    "keywords": [
-        "micro services",
-        "micro frontend",
-        "components",
-        "podium"
-    ],
-    "repository": {
-        "type": "git",
-        "url": "git@github.com:podium-lib/utils.git"
-    },
-    "bugs": {
-        "url": "https://github.com/podium-lib/issues"
-    },
-    "homepage": "https://podium-lib.io/",
-    "files": [
-        "package.json",
-        "CHANGELOG.md",
-        "index.d.ts",
-        "README.md",
-        "LICENSE",
-        "dist",
-        "lib"
-    ],
-    "main": "./lib/main.js",
-    "types": "index.d.ts",
-    "scripts": {
-        "lint": "eslint .",
-        "lint:fix": "eslint --fix .",
-        "test": "tap --no-check-coverage",
-        "test:snapshots:update": "tap --snapshot",
-        "bench": "node benchmark/benchmark.js"
-    },
-    "devDependencies": {
-        "@semantic-release/git": "10.0.1",
-        "@semantic-release/changelog": "6.0.3",
-        "@semantic-release/commit-analyzer": "11.1.0",
-        "@semantic-release/github": "9.2.6",
-        "@semantic-release/npm": "11.0.3",
-        "@semantic-release/release-notes-generator": "12.1.0",
-        "benchmark": "2.1.4",
-        "eslint": "8.57.0",
-        "eslint-config-airbnb-base": "15.0.0",
-        "eslint-config-prettier": "9.1.0",
-        "eslint-plugin-import": "2.29.1",
-        "eslint-plugin-prettier": "5.1.3",
-        "@babel/eslint-parser": "7.24.1",
-        "tap": "16.3.10",
-        "prettier": "3.2.5",
-        "@podium/schemas": "5.0.0"
-    },
-    "dependencies": {
-        "camelcase": "8.0.0"
-    }
+  "name": "@podium/utils",
+  "version": "5.0.3",
+  "description": "Common generic utility methods shared by @podium modules.",
+  "type": "module",
+  "license": "MIT",
+  "keywords": [
+    "micro services",
+    "micro frontend",
+    "components",
+    "podium"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "git@github.com:podium-lib/utils.git"
+  },
+  "bugs": {
+    "url": "https://github.com/podium-lib/issues"
+  },
+  "homepage": "https://podium-lib.io/",
+  "files": [
+    "package.json",
+    "CHANGELOG.md",
+    "README.md",
+    "LICENSE",
+    "dist",
+    "lib",
+    "types.d.ts"
+  ],
+  "main": "./lib/main.js",
+  "types": "./types/main.d.ts",
+  "scripts": {
+    "bench": "node benchmark/benchmark.js",
+    "lint": "eslint .",
+    "lint:fix": "eslint --fix .",
+    "test": "tap --no-check-coverage && tsc --project tsconfig.test.json",
+    "test:snapshots:update": "tap --snapshot",
+    "types": "tsc --declaration --emitDeclarationOnly"
+  },
+  "devDependencies": {
+    "@babel/eslint-parser": "7.24.1",
+    "@podium/schemas": "5.0.0",
+    "@semantic-release/changelog": "6.0.3",
+    "@semantic-release/commit-analyzer": "11.1.0",
+    "@semantic-release/git": "10.0.1",
+    "@semantic-release/github": "9.2.6",
+    "@semantic-release/npm": "11.0.3",
+    "@semantic-release/release-notes-generator": "12.1.0",
+    "@types/node": "20.12.7",
+    "benchmark": "2.1.4",
+    "eslint": "8.57.0",
+    "eslint-config-airbnb-base": "15.0.0",
+    "eslint-config-prettier": "9.1.0",
+    "eslint-plugin-import": "2.29.1",
+    "eslint-plugin-prettier": "5.1.3",
+    "prettier": "3.2.5",
+    "tap": "16.3.10",
+    "typescript": "5.4.5"
+  },
+  "dependencies": {
+    "camelcase": "8.0.0"
+  }
 }
