+++ tests/asset-js.test.js
