+++ tests/html-document.test.js
@@ -2,7 +2,6 @@ import tap from 'tap';
 import HttpIncoming from '../lib/http-incoming.js';
 import { document } from '../lib/html-document.js';
 
-
 const SIMPLE_REQ = {
     headers: {},
 };
@@ -33,28 +32,31 @@ tap.test('.document() - arguments given', (t) => {
     const head = 'this goes in the head section';
     const body = 'this goes in the body section';
 
-    const result = document(incoming, body, head);    
+    const result = document(incoming, body, head);
     t.matchSnapshot(result, 'should render template using values given');
     t.end();
 });
 
-tap.test('.document() - arguments given - handles v4 js and css syntax', (t) => {
-    const incoming = new HttpIncoming(SIMPLE_REQ, SIMPLE_RES);
-    incoming.css = [
-        { value: 'http://somecssurl1.com', type: 'text/css' },
-        { value: 'http://somecssurl2.com', type: 'text/css' },
-        { value: 'http://somecssurl3.com', type: 'text/css' },
-    ];
-    incoming.js = [
-        { value: 'http://somejsurl1.com', type: 'default' },
-        { value: 'http://somejsurl2.com', type: 'default' },
-        { value: 'http://somejsurl3.com', type: 'default' },
-    ];
+tap.test(
+    '.document() - arguments given - handles v4 js and css syntax',
+    (t) => {
+        const incoming = new HttpIncoming(SIMPLE_REQ, SIMPLE_RES);
+        incoming.css = [
+            { value: 'http://somecssurl1.com', type: 'text/css' },
+            { value: 'http://somecssurl2.com', type: 'text/css' },
+            { value: 'http://somecssurl3.com', type: 'text/css' },
+        ];
+        incoming.js = [
+            { value: 'http://somejsurl1.com', type: 'default' },
+            { value: 'http://somejsurl2.com', type: 'default' },
+            { value: 'http://somejsurl3.com', type: 'default' },
+        ];
 
-    const result = document(incoming);
-    t.matchSnapshot(result, 'should render template using values given');
-    t.end();
-});
+        const result = document(incoming);
+        t.matchSnapshot(result, 'should render template using values given');
+        t.end();
+    },
+);
 
 tap.test('.document() - js "type" is "esm"', (t) => {
     const incoming = new HttpIncoming(SIMPLE_REQ, SIMPLE_RES);
@@ -74,19 +76,34 @@ tap.test('.document() - js "type" is "esm"', (t) => {
     t.end();
 });
 
-tap.test('.document() - "type" is "module", "strategy" is set - should place assets based on strategy', (t) => {
-    const incoming = new HttpIncoming(SIMPLE_REQ, SIMPLE_RES);
-    incoming.css = [{ value: 'http://somecssurl1.com', type: 'text/css' }];
-    incoming.js = [
-        { value: 'http://somejsurl1.com/lazy', type: 'module', strategy: 'lazy' },
-        { value: 'http://somejsurl2.com/before', type: 'module', strategy: 'beforeInteractive' },
-        { value: 'http://somejsurl3.com/after', type: 'module', strategy: 'afterInteractive' },
-    ];
+tap.test(
+    '.document() - "type" is "module", "strategy" is set - should place assets based on strategy',
+    (t) => {
+        const incoming = new HttpIncoming(SIMPLE_REQ, SIMPLE_RES);
+        incoming.css = [{ value: 'http://somecssurl1.com', type: 'text/css' }];
+        incoming.js = [
+            {
+                value: 'http://somejsurl1.com/lazy',
+                type: 'module',
+                strategy: 'lazy',
+            },
+            {
+                value: 'http://somejsurl2.com/before',
+                type: 'module',
+                strategy: 'beforeInteractive',
+            },
+            {
+                value: 'http://somejsurl3.com/after',
+                type: 'module',
+                strategy: 'afterInteractive',
+            },
+        ];
 
-    const result = document(incoming);
-    t.matchSnapshot(result);
-    t.end();
-});
+        const result = document(incoming);
+        t.matchSnapshot(result);
+        t.end();
+    },
+);
 
 tap.test('.document() - lang tag priority - params has priority', (t) => {
     const incoming = new HttpIncoming(SIMPLE_REQ, SIMPLE_RES, { locale: 'sv' });
@@ -106,7 +123,7 @@ tap.test('.document() - lang tag priority - context has priority', (t) => {
 tap.test('.document() - lang tag priority - view has priority', (t) => {
     const incoming = new HttpIncoming(SIMPLE_REQ, SIMPLE_RES, { locale: 'sv' });
     incoming.context = { locale: 'en-NZ' };
-    incoming.view = { locale: 'nb' }
+    incoming.view = { locale: 'nb' };
     const result = document(incoming, '');
     t.match(result, 'lang="nb"', 'should render lang tag nb');
     t.end();
